Conflicts:
	src/engine/server.h
	src/engine/server/server.cpp
	src/engine/server/server.h
	src/game/server/entities/character.cpp
	src/game/server/eventhandler.h
	src/game/server/gamecontext.h
	src/game/server/gameworld.cpp
	src/game/server/player.cpp
	src/game/variables.h

+++ src/game/server/eventhandler.h
@@ -13,7 +13,7 @@ class CEventHandler
 	int m_aTypes[MAX_EVENTS]; // TODO: remove some of these arrays
 	int m_aOffsets[MAX_EVENTS];
 	int m_aSizes[MAX_EVENTS];
-	int m_aClientMasks[MAX_EVENTS];
+	int64_t m_aClientMasks[MAX_EVENTS];
 	char m_aData[MAX_DATASIZE];
 
 	class CGameContext *m_pGameServer;
